register

+++ database/migrations/2020_04_02_024420_create_departamento_table.php
@@ -18,8 +18,8 @@ class CreateDepartamentoTable extends Migration
 			$table->integer('co_departamento')->unique();
 			$table->string('nb_departamento', 30)->unique();
 			$table->integer('id_pais');
-			$table->integer('nu_latitud')->nullable();
-			$table->integer('nu_longitud')->nullable();
+			$table->float('nu_latitud', 2, 7)->nullable();
+			$table->float('nu_longitud', 2, 7)->nullable();
 			$table->string('tx_observaciones', 100)->nullable();
 			$table->integer('id_status');
 			$table->integer('id_usuario');
