+++ database/migrations/2020_04_02_024501_create_comuna_table.php
@@ -18,8 +18,8 @@ class CreateComunaTable extends Migration
 			$table->integer('co_comuna')->unique();
 			$table->string('nb_comuna', 30);
 			$table->integer('id_zona');
-			$table->integer('nu_latitud')->nullable();
-			$table->integer('nu_longitud')->nullable();
+			$table->float('nu_latitud', 2, 7)->nullable();
+			$table->float('nu_longitud', 2, 7)->nullable();
 			$table->string('tx_observaciones', 100)->nullable();
 			$table->integer('id_status');
 			$table->integer('id_usuario');
