+++ database/migrations/2020_04_02_024512_create_barrio_table.php
@@ -18,8 +18,8 @@ class CreateBarrioTable extends Migration
 			$table->integer('co_barrio')->unique();
 			$table->string('nb_barrio', 30);
 			$table->integer('id_comuna');
-			$table->integer('nu_latitud')->nullable();
-			$table->integer('nu_longitud')->nullable();
+			$table->float('nu_latitud', 2, 7)->nullable();
+			$table->float('nu_longitud', 2, 7)->nullable();
 			$table->string('tx_observaciones', 100)->nullable();
 			$table->integer('id_status');
 			$table->integer('id_usuario');
